Move ImmutableStack types out
