+++ src/System.Collections.Immutable/src/System.Collections.Immutable.csproj
@@ -83,6 +83,8 @@
     <Compile Include="System\Collections\Immutable\ImmutableSortedSet_1.Node.cs" />
     <Compile Include="System\Collections\Immutable\ImmutableStack.cs" />
     <Compile Include="System\Collections\Immutable\ImmutableStack_1.cs" />
+    <Compile Include="System\Collections\Immutable\ImmutableStack_1.DebuggerProxy.cs" />
+    <Compile Include="System\Collections\Immutable\ImmutableStack_1.Enumerator.cs" />
     <Compile Include="System\Collections\Immutable\KeysOrValuesCollectionAccessor.cs" />
     <Compile Include="System\Collections\Immutable\RefAsValueType.cs" />
     <Compile Include="System\Collections\Immutable\SecureObjectPool.cs" />
