+++ src/System.Collections.Immutable/src/System/Collections/Immutable/ImmutableStack_1.DebuggerProxy.cs
@@ -1,0 +1,52 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.Diagnostics;
+using System.Linq;
+
+namespace System.Collections.Immutable
+{
+    /// <summary>
+    /// A simple view of the immutable collection that the debugger can show to the developer.
+    /// </summary>
+    internal class ImmutableStackDebuggerProxy<T>
+    {
+        /// <summary>
+        /// The collection to be enumerated.
+        /// </summary>
+        private readonly ImmutableStack<T> _stack;
+
+        /// <summary>
+        /// The simple view of the collection.
+        /// </summary>
+        private T[] _contents;
+
+        /// <summary>   
+        /// Initializes a new instance of the <see cref="ImmutableStackDebuggerProxy{T}"/> class.
+        /// </summary>
+        /// <param name="stack">The collection to display in the debugger</param>
+        public ImmutableStackDebuggerProxy(ImmutableStack<T> stack)
+        {
+            Requires.NotNull(stack, nameof(stack));
+            _stack = stack;
+        }
+
+        /// <summary>
+        /// Gets a simple debugger-viewable collection.
+        /// </summary>
+        [DebuggerBrowsable(DebuggerBrowsableState.RootHidden)]
+        public T[] Contents
+        {
+            get
+            {
+                if (_contents == null)
+                {
+                    _contents = _stack.ToArray();
+                }
+
+                return _contents;
+            }
+        }
+    }
+}
