+++ src/System.Collections.Immutable/src/System/Collections/Immutable/ImmutableStack_1.cs
@@ -3,11 +3,9 @@
 // See the LICENSE file in the project root for more information.
 
 using System.Collections.Generic;
-using System.ComponentModel;
 using System.Diagnostics;
 using System.Diagnostics.CodeAnalysis;
 using System.Diagnostics.Contracts;
-using System.Linq;
 
 namespace System.Collections.Immutable
 {
@@ -19,7 +17,7 @@ namespace System.Collections.Immutable
     [DebuggerTypeProxy(typeof(ImmutableStackDebuggerProxy<>))]
     [SuppressMessage("Microsoft.Naming", "CA1710:IdentifiersShouldHaveCorrectSuffix", Justification = "Ignored")]
     [SuppressMessage("Microsoft.Naming", "CA1711:IdentifiersShouldNotHaveIncorrectSuffix", Justification = "Ignored")]
-    public sealed class ImmutableStack<T> : IImmutableStack<T>
+    public sealed partial class ImmutableStack<T> : IImmutableStack<T>
     {
         /// <summary>
         /// The singleton empty stack.
@@ -241,222 +239,6 @@ namespace System.Collections.Immutable
             }
 
             return r;
-        }
-
-        /// <summary>
-        /// Enumerates a stack with no memory allocations.
-        /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Advanced)]
-        public struct Enumerator
-        {
-            /// <summary>
-            /// The original stack being enumerated.
-            /// </summary>
-            private readonly ImmutableStack<T> _originalStack;
-
-            /// <summary>
-            /// The remaining stack not yet enumerated.
-            /// </summary>
-            private ImmutableStack<T> _remainingStack;
-
-            /// <summary>
-            /// Initializes a new instance of the <see cref="Enumerator"/> struct.
-            /// </summary>
-            /// <param name="stack">The stack to enumerator.</param>
-            internal Enumerator(ImmutableStack<T> stack)
-            {
-                Requires.NotNull(stack, nameof(stack));
-                _originalStack = stack;
-                _remainingStack = null;
-            }
-
-            /// <summary>
-            /// Gets the current element.
-            /// </summary>
-            public T Current
-            {
-                get
-                {
-                    if (_remainingStack == null || _remainingStack.IsEmpty)
-                    {
-                        throw new InvalidOperationException();
-                    }
-                    else
-                    {
-                        return _remainingStack.Peek();
-                    }
-                }
-            }
-
-            /// <summary>
-            /// Moves to the first or next element.
-            /// </summary>
-            /// <returns>A value indicating whether there are any more elements.</returns>
-            public bool MoveNext()
-            {
-                if (_remainingStack == null)
-                {
-                    // initial move
-                    _remainingStack = _originalStack;
-                }
-                else if (!_remainingStack.IsEmpty)
-                {
-                    _remainingStack = _remainingStack.Pop();
-                }
-
-                return !_remainingStack.IsEmpty;
-            }
-        }
-
-        /// <summary>
-        /// Enumerates a stack with no memory allocations.
-        /// </summary>
-        private class EnumeratorObject : IEnumerator<T>
-        {
-            /// <summary>
-            /// The original stack being enumerated.
-            /// </summary>
-            private readonly ImmutableStack<T> _originalStack;
-
-            /// <summary>
-            /// The remaining stack not yet enumerated.
-            /// </summary>
-            private ImmutableStack<T> _remainingStack;
-
-            /// <summary>
-            /// A flag indicating whether this enumerator has been disposed.
-            /// </summary>
-            private bool _disposed;
-
-            /// <summary>
-            /// Initializes a new instance of the <see cref="EnumeratorObject"/> class.
-            /// </summary>
-            /// <param name="stack">The stack to enumerator.</param>
-            internal EnumeratorObject(ImmutableStack<T> stack)
-            {
-                Requires.NotNull(stack, nameof(stack));
-                _originalStack = stack;
-            }
-
-            /// <summary>
-            /// Gets the current element.
-            /// </summary>
-            public T Current
-            {
-                get
-                {
-                    this.ThrowIfDisposed();
-                    if (_remainingStack == null || _remainingStack.IsEmpty)
-                    {
-                        throw new InvalidOperationException();
-                    }
-                    else
-                    {
-                        return _remainingStack.Peek();
-                    }
-                }
-            }
-
-            /// <summary>
-            /// Gets the current element.
-            /// </summary>
-            object IEnumerator.Current
-            {
-                get { return this.Current; }
-            }
-
-            /// <summary>
-            /// Moves to the first or next element.
-            /// </summary>
-            /// <returns>A value indicating whether there are any more elements.</returns>
-            public bool MoveNext()
-            {
-                this.ThrowIfDisposed();
-
-                if (_remainingStack == null)
-                {
-                    // initial move
-                    _remainingStack = _originalStack;
-                }
-                else if (!_remainingStack.IsEmpty)
-                {
-                    _remainingStack = _remainingStack.Pop();
-                }
-
-                return !_remainingStack.IsEmpty;
-            }
-
-            /// <summary>
-            /// Resets the position to just before the first element in the list.
-            /// </summary>
-            public void Reset()
-            {
-                this.ThrowIfDisposed();
-                _remainingStack = null;
-            }
-
-            /// <summary>
-            /// Disposes this instance.
-            /// </summary>
-            public void Dispose()
-            {
-                _disposed = true;
-            }
-
-            /// <summary>
-            /// Throws an <see cref="ObjectDisposedException"/> if this 
-            /// enumerator has already been disposed.
-            /// </summary>
-            private void ThrowIfDisposed()
-            {
-                if (_disposed)
-                {
-                    Requires.FailObjectDisposed(this);
-                }
-            }
-        }
-    }
-
-    /// <summary>
-    /// A simple view of the immutable collection that the debugger can show to the developer.
-    /// </summary>
-    internal class ImmutableStackDebuggerProxy<T>
-    {
-        /// <summary>
-        /// The collection to be enumerated.
-        /// </summary>
-        private readonly ImmutableStack<T> _stack;
-
-        /// <summary>
-        /// The simple view of the collection.
-        /// </summary>
-        private T[] _contents;
-
-        /// <summary>   
-        /// Initializes a new instance of the <see cref="ImmutableStackDebuggerProxy{T}"/> class.
-        /// </summary>
-        /// <param name="stack">The collection to display in the debugger</param>
-        public ImmutableStackDebuggerProxy(ImmutableStack<T> stack)
-        {
-            Requires.NotNull(stack, nameof(stack));
-            _stack = stack;
-        }
-
-        /// <summary>
-        /// Gets a simple debugger-viewable collection.
-        /// </summary>
-        [DebuggerBrowsable(DebuggerBrowsableState.RootHidden)]
-        public T[] Contents
-        {
-            get
-            {
-                if (_contents == null)
-                {
-                    _contents = _stack.ToArray();
-                }
-
-                return _contents;
-            }
         }
     }
 }
